break out utils into cmip6 utils
