+++ aneris/__init__.py
@@ -1,7 +1,7 @@
 from aneris._io import *
 from aneris.harmonize import *
 from aneris.utils import *
-from aneris.cmip6 import *
+from aneris.cmip6 import driver, cmip6_utils
 
 # _version is generated by setuptools_scm upon installing with pip
 from aneris._version import __version__, __version_tuple__  # noqa: F401
