+++ aneris/cmip6/__init__.py
@@ -1,1 +1,2 @@
 from .driver import *
+from .cmip6_utils import *
